Restructured app module into separate submodules for better organisation
